+++ lib/src/winrt/ICalendar.dart
@@ -23,10 +23,10 @@ import '../structs.g.dart';
 import '../utils.dart';
 
 import '../api-ms-win-core-winrt-string-l1-1-0.dart';
-import '../winrt/winrt_helpers.dart';
+import '../winrt_helpers.dart';
 import '../types.dart';
 
-import 'IInspectable.dart';
+import '../com/IInspectable.dart';
 import 'IVectorView`1.dart';
 
 /// @nodoc
