+++ lib/src/winrt/ToastNotification.dart
@@ -8,20 +8,18 @@ import 'package:ffi/ffi.dart';
 
 import '../api-ms-win-core-winrt-l1-1-0.dart';
 import '../api-ms-win-core-winrt-string-l1-1-0.dart';
+import '../com/IInspectable.dart';
 import '../combase.dart';
 import '../constants.dart';
 import '../exceptions.dart';
 import '../guid.dart';
-
 import '../macros.dart';
 import '../ole32.dart';
 import '../structs.dart';
 import '../structs.g.dart';
 import '../utils.dart';
-import '../winrt/winrt_constants.dart';
-import '../winrt/winrt_helpers.dart';
-
-import 'IInspectable.dart';
+import '../winrt_constants.dart';
+import '../winrt_helpers.dart';
 import 'IToastNotificationFactory.dart';
 
 const _className = 'Windows.UI.Notifications.ToastNotification';
