+++ lib/src/winrt_constants.dart
