+++ lib/win32.dart
@@ -106,7 +106,6 @@ export 'src/extensions/unpack_utf16.dart';
 export 'src/advapi32.dart';
 export 'src/bluetoothapis.dart';
 export 'src/bthprops.dart';
-export 'src/combase.dart';
 export 'src/comctl32.dart';
 export 'src/comdlg32.dart';
 export 'src/dbghelp.dart';
@@ -141,11 +140,12 @@ export 'src/api-ms-win-core-winrt-string-l1-1-0.dart';
 export 'src/api-ms-win-ro-typeresolution-l1-1-0.dart';
 export 'src/api-ms-win-wsl-api-l1-1-0.dart';
 
-// WinRT foundational exports
-export 'src/winrt/winrt_constants.dart';
-export 'src/winrt/winrt_helpers.dart';
+// COM and Windows Runtime foundational exports
+export 'src/combase.dart';
+export 'src/winrt_constants.dart';
+export 'src/winrt_helpers.dart';
 
-// COM and Windows Runtime interfaces
+// COM interfaces
 export 'src/com/IApplicationActivationManager.dart';
 export 'src/com/IApplicationActivationManager.dart';
 export 'src/com/IAppxFactory.dart';
@@ -167,8 +167,6 @@ export 'src/com/IAppxManifestReader5.dart';
 export 'src/com/IAppxManifestReader6.dart';
 export 'src/com/IAppxManifestReader7.dart';
 export 'src/com/IAppxPackageReader.dart';
-export 'src/com/IAsyncAction.dart';
-export 'src/com/IAsyncInfo.dart';
 export 'src/com/IAudioCaptureClient.dart';
 export 'src/com/IAudioClient.dart';
 export 'src/com/IAudioClock.dart';
@@ -177,7 +175,6 @@ export 'src/com/IAudioSessionControl.dart';
 export 'src/com/IAudioSessionManager.dart';
 export 'src/com/IAudioStreamVolume.dart';
 export 'src/com/IBindCtx.dart';
-export 'src/com/ICalendar.dart';
 export 'src/com/IChannelAudioVolume.dart';
 export 'src/com/IClassFactory.dart';
 export 'src/com/IConnectionPoint.dart';
@@ -199,9 +196,7 @@ export 'src/com/IFileDialog2.dart';
 export 'src/com/IFileDialogCustomize.dart';
 export 'src/com/IFileIsInUse.dart';
 export 'src/com/IFileOpenDialog.dart';
-export 'src/com/IFileOpenPicker.dart';
 export 'src/com/IFileSaveDialog.dart';
-export 'src/com/IHostName.dart';
 export 'src/com/IInspectable.dart';
 export 'src/com/IKnownFolder.dart';
 export 'src/com/IKnownFolderManager.dart';
@@ -211,14 +206,12 @@ export 'src/com/IModalWindow.dart';
 export 'src/com/IMoniker.dart';
 export 'src/com/INetwork.dart';
 export 'src/com/INetworkConnection.dart';
-export 'src/com/INetworkInformationStatics.dart';
 export 'src/com/INetworkListManager.dart';
 export 'src/com/INetworkListManagerEvents.dart';
 export 'src/com/IPersist.dart';
 export 'src/com/IPersistFile.dart';
 export 'src/com/IPersistMemory.dart';
 export 'src/com/IPersistStream.dart';
-export 'src/com/IPropertyValue.dart';
 export 'src/com/IProvideClassInfo.dart';
 export 'src/com/IRunningObjectTable.dart';
 export 'src/com/ISequentialStream.dart';
@@ -233,11 +226,11 @@ export 'src/com/IShellLinkDataList.dart';
 export 'src/com/IShellLinkDual.dart';
 export 'src/com/IShellService.dart';
 export 'src/com/ISimpleAudioVolume.dart';
-export 'src/com/ISpeechObjectToken.dart';
-export 'src/com/ISpeechObjectTokens.dart';
 export 'src/com/ISpEventSource.dart';
 export 'src/com/ISpNotifySource.dart';
 export 'src/com/ISpVoice.dart';
+export 'src/com/ISpeechObjectToken.dart';
+export 'src/com/ISpeechObjectTokens.dart';
 export 'src/com/ISpellChecker.dart';
 export 'src/com/ISpellChecker2.dart';
 export 'src/com/ISpellCheckerChangedEventHandler.dart';
@@ -245,14 +238,10 @@ export 'src/com/ISpellCheckerFactory.dart';
 export 'src/com/ISpellingError.dart';
 export 'src/com/IStream.dart';
 export 'src/com/ISupportErrorInfo.dart';
-export 'src/com/IToastNotificationFactory.dart';
-export 'src/com/IToastNotificationManagerStatics.dart';
 export 'src/com/ITypeInfo.dart';
 export 'src/com/IUnknown.dart';
 export 'src/com/IUri.dart';
 export 'src/com/IUserDataPathsStatics.dart';
-export 'src/com/IVector`1.dart';
-export 'src/com/IVectorView`1.dart';
 export 'src/com/IVirtualDesktopManager.dart';
 export 'src/com/IWbemClassObject.dart';
 export 'src/com/IWbemConfigureRefresher.dart';
@@ -262,5 +251,18 @@ export 'src/com/IWbemLocator.dart';
 export 'src/com/IWbemObjectAccess.dart';
 export 'src/com/IWbemRefresher.dart';
 export 'src/com/IWbemServices.dart';
-export 'src/com/ToastNotification.dart';
-export 'src/com/UserDataPaths.dart';
+
+// Windows Runtime interfaces
+export 'src/winrt/IAsyncAction.dart';
+export 'src/winrt/IAsyncInfo.dart';
+export 'src/winrt/ICalendar.dart';
+export 'src/winrt/IFileOpenPicker.dart';
+export 'src/winrt/IHostName.dart';
+export 'src/winrt/INetworkInformationStatics.dart';
+export 'src/winrt/IPropertyValue.dart';
+export 'src/winrt/IToastNotificationFactory.dart';
+export 'src/winrt/IToastNotificationManagerStatics.dart';
+export 'src/winrt/IVectorView`1.dart';
+export 'src/winrt/IVector`1.dart';
+export 'src/winrt/ToastNotification.dart';
+export 'src/winrt/UserDataPaths.dart';
